Initial attempts at updating and merging Bay12 code while retaining Exodus map and functions.

# Conflicts:
#	baystation12.dme
#	code/ATMOSPHERICS/components/trinary_devices/filter.dm
#	code/ATMOSPHERICS/components/unary/vent_scrubber.dm
#	code/__HELPERS/text.dm
#	code/datums/supplypacks.dm
#	code/game/area/Space Station 13 areas.dm
#	code/game/gamemodes/gameticker.dm
#	code/game/jobs/job_controller.dm
#	code/game/machinery/alarm.dm
#	code/game/machinery/atmoalter/scrubber.dm
#	code/game/machinery/autolathe.dm
#	code/game/machinery/computer/supply.dm
#	code/game/machinery/computer3/computer.dm
#	code/game/machinery/doors/brigdoors.dm
#	code/game/machinery/doors/door.dm
#	code/game/machinery/doors/firedoor.dm
#	code/game/machinery/telecomms/presets.dm
#	code/game/mecha/combat/durand.dm
#	code/game/mecha/combat/gygax.dm
#	code/game/mecha/combat/marauder.dm
#	code/game/mecha/combat/phazon.dm
#	code/game/mecha/equipment/weapons/weapons.dm
#	code/game/mecha/medical/odysseus.dm
#	code/game/mecha/working/ripley.dm
#	code/game/objects/effects/landmarks.dm
#	code/game/objects/explosion.dm
#	code/game/objects/items/devices/flashlight.dm
#	code/game/objects/items/weapons/manuals.dm
#	code/game/objects/items/weapons/power_cells.dm
#	code/game/objects/items/weapons/storage/boxes.dm
#	code/game/objects/items/weapons/tanks/tank_types.dm
#	code/game/objects/structures/crates_lockers/closets/secure/engineering.dm
#	code/game/objects/structures/crates_lockers/closets/secure/medical.dm
#	code/game/objects/structures/crates_lockers/closets/secure/security.dm
#	code/game/supplyshuttle.dm
#	code/game/turfs/simulated/floor_types.dm
#	code/game/turfs/space/space.dm
#	code/global.dm
#	code/hub.dm
#	code/modules/client/preferences_spawnpoints.dm
#	code/modules/clothing/head/helmet.dm
#	code/modules/clothing/head/jobs.dm
#	code/modules/clothing/spacesuits/alien.dm
#	code/modules/clothing/spacesuits/captain.dm
#	code/modules/clothing/under/accessories/storage.dm
#	code/modules/mining/mine_items.dm
#	code/modules/mining/mine_turfs.dm
#	code/modules/mob/living/carbon/human/human_movement.dm
#	code/modules/mob/living/carbon/human/inventory.dm
#	code/modules/mob/living/carbon/human/life.dm
#	code/modules/mob/living/carbon/human/species/outsider/vox.dm
#	code/modules/mob/living/carbon/human/species/station/station.dm
#	code/modules/mob/living/carbon/human/species/xenomorphs/alien_embryo.dm
#	code/modules/mob/new_player/new_player.dm
#	code/modules/power/apc.dm
#	code/modules/power/cell.dm
#	code/modules/power/generator.dm
#	code/modules/power/lighting.dm
#	code/modules/projectiles/guns/energy.dm
#	code/modules/projectiles/guns/energy/nuclear.dm
#	code/modules/projectiles/guns/projectile/pistol.dm
#	code/modules/projectiles/projectile/beams.dm
#	code/modules/random_map/mining_distribution.dm
#	code/modules/random_map/random_map.dm
#	code/modules/reagents/Chemistry-Recipes.dm
#	code/modules/reagents/dispenser/_defines.dm
#	code/modules/reagents/reagent_containers/glass.dm
#	code/modules/reagents/reagent_dispenser.dm
#	code/modules/research/designs.dm
#	code/modules/research/xenoarchaeology/misc.dm
#	code/modules/shuttles/shuttle_emergency.dm
#	icons/mob/items/lefthand_guns.dmi
#	icons/mob/items/righthand_guns.dmi
#	icons/obj/gun.dmi
#	icons/obj/mining.dmi
#	icons/obj/projectiles.dmi
#	icons/obj/tank.dmi
#	icons/obj/watercloset.dmi
#	maps/exodus-1.dmm
#	maps/exodus-5.dmm

+++ code/game/machinery/kitchen/smartfridge.dm
@@ -1,21 +1,22 @@
+
 /* SmartFridge.  Much todo
 */
 /obj/machinery/smartfridge
 	name = "\improper SmartFridge"
 	icon = 'icons/obj/vending.dmi'
 	icon_state = "smartfridge"
-	layer = 2.9
+	layer = BELOW_OBJ_LAYER
 	density = 1
 	anchored = 1
-	use_power = 1
 	idle_power_usage = 5
 	active_power_usage = 100
-	flags = NOREACT
+	atom_flags = ATOM_FLAG_NO_TEMP_CHANGE | ATOM_FLAG_NO_REACT
 	var/global/max_n_of_items = 999 // Sorry but the BYOND infinite loop detector doesn't look things over 1000.
 	var/icon_on = "smartfridge"
 	var/icon_off = "smartfridge-off"
 	var/icon_panel = "smartfridge-panel"
-	var/item_quants = list()
+	var/list/item_records = list()
+	var/datum/stored_items/currently_vending = null	//What we're putting out of the machine.
 	var/seconds_electrified = 0;
 	var/shoot_inventory = 0
 	var/locked = 0
@@ -35,7 +36,11 @@
 
 /obj/machinery/smartfridge/Destroy()
 	qdel(wires)
-	..()
+	wires = null
+	for(var/datum/stored_items/S in item_records)
+		qdel(S)
+	item_records = null
+	return ..()
 
 /obj/machinery/smartfridge/proc/accept_check(var/obj/item/O as obj)
 	if(istype(O,/obj/item/weapon/reagent_containers/food/snacks/grown/) || istype(O,/obj/item/seeds/))
@@ -57,7 +62,7 @@
 
 /obj/machinery/smartfridge/secure/extract
 	name = "\improper Slime Extract Storage"
-	desc = "A refrigerated storage unit for slime extracts"
+	desc = "A refrigerated storage unit for slime extracts."
 	req_access = list(access_research)
 
 /obj/machinery/smartfridge/secure/extract/accept_check(var/obj/item/O as obj)
@@ -118,9 +123,23 @@
 	if(istype(O,/obj/item/weapon/reagent_containers/glass) || istype(O,/obj/item/weapon/reagent_containers/food/drinks) || istype(O,/obj/item/weapon/reagent_containers/food/condiment))
 		return 1
 
+/obj/machinery/smartfridge/foods
+	name = "\improper Hot Foods Display"
+	desc = "A heated storage unit for piping hot meals."
+	icon_state = "smartfridge_food"
+	icon_on = "smartfridge_food"
+	icon_off = "smartfridge_food-off"
+
+/obj/machinery/smartfridge/foods/accept_check(var/obj/item/O as obj)
+	if(istype(O,/obj/item/weapon/reagent_containers/food/snacks) || istype(O,/obj/item/weapon/material/kitchen/utensil))
+		return 1
+
 /obj/machinery/smartfridge/drying_rack
 	name = "\improper Drying Rack"
 	desc = "A machine for drying plants."
+	icon_state = "drying_rack"
+	icon_on = "drying_rack_on"
+	icon_off = "drying_rack"
 
 /obj/machinery/smartfridge/drying_rack/accept_check(var/obj/item/O as obj)
 	if(istype(O, /obj/item/weapon/reagent_containers/food/snacks/))
@@ -129,29 +148,41 @@
 			return 1
 	return 0
 
-/obj/machinery/smartfridge/drying_rack/process()
+/obj/machinery/smartfridge/drying_rack/Process()
 	..()
-	if (contents.len)
+	if(inoperable())
+		return
+	if(contents.len)
 		dry()
+		update_icon()
+
+/obj/machinery/smartfridge/drying_rack/on_update_icon()
+	overlays.Cut()
+	if(inoperable())
+		icon_state = icon_off
+	else
+		icon_state = icon_on
+	if(contents.len)
+		overlays += "drying_rack_filled"
+		if(!inoperable())
+			overlays += "drying_rack_drying"
 
 /obj/machinery/smartfridge/drying_rack/proc/dry()
-	for(var/obj/item/weapon/reagent_containers/food/snacks/S in contents)
-		if(S.dry) continue
-		if(S.dried_type == S.type)
-			S.dry = 1
-			item_quants[S.name]--
-			S.name = "dried [S.name]"
-			S.color = "#AAAAAA"
-			S.loc = loc
-		else
-			var/D = S.dried_type
-			new D(loc)
-			item_quants[S.name]--
-			qdel(S)
-		return
-	return
+	for(var/datum/stored_items/I in item_records)
+		for(var/obj/item/weapon/reagent_containers/food/snacks/S in I.instances)
+			if(S.dry || !I.get_specific_product(get_turf(src), S)) continue
+			if(S.dried_type == S.type)
+				S.dry = 1
+				S.SetName("dried [S.name]")
+				S.color = "#a38463"
+				stock_item(S)
+			else
+				var/D = S.dried_type
+				new D(get_turf(src))
+				qdel(S)
+			return
 
-/obj/machinery/smartfridge/process()
+/obj/machinery/smartfridge/Process()
 	if(stat & (BROKEN|NOPOWER))
 		return
 	if(src.seconds_electrified > 0)
@@ -159,13 +190,7 @@
 	if(src.shoot_inventory && prob(2))
 		src.throw_item()
 
-/obj/machinery/smartfridge/power_change()
-	var/old_stat = stat
-	..()
-	if(old_stat != stat)
-		update_icon()
-
-/obj/machinery/smartfridge/update_icon()
+/obj/machinery/smartfridge/on_update_icon()
 	if(stat & (BROKEN|NOPOWER))
 		icon_state = icon_off
 	else
@@ -176,74 +201,68 @@
 ********************/
 
 /obj/machinery/smartfridge/attackby(var/obj/item/O as obj, var/mob/user as mob)
-	if(istype(O, /obj/item/weapon/screwdriver))
+	if(isScrewdriver(O))
 		panel_open = !panel_open
 		user.visible_message("[user] [panel_open ? "opens" : "closes"] the maintenance panel of \the [src].", "You [panel_open ? "open" : "close"] the maintenance panel of \the [src].")
 		overlays.Cut()
 		if(panel_open)
 			overlays += image(icon, icon_panel)
-		nanomanager.update_uis(src)
+		SSnano.update_uis(src)
 		return
 
-	if(istype(O, /obj/item/device/multitool)||istype(O, /obj/item/weapon/wirecutters))
+	if(isMultitool(O) || isWirecutter(O))
 		if(panel_open)
 			attack_hand(user)
 		return
 
 	if(stat & NOPOWER)
-		user << "<span class='notice'>\The [src] is unpowered and useless.</span>"
+		to_chat(user, "<span class='notice'>\The [src] is unpowered and useless.</span>")
 		return
 
 	if(accept_check(O))
-		if(contents.len >= max_n_of_items)
-			user << "<span class='notice'>\The [src] is full.</span>"
-			return 1
-		else
-			user.remove_from_mob(O)
-			O.loc = src
-			if(item_quants[O.name])
-				item_quants[O.name]++
-			else
-				item_quants[O.name] = 1
-			user.visible_message("<span class='notice'>[user] has added \the [O] to \the [src].</span>", "<span class='notice'>You add \the [O] to \the [src].</span>")
+		if(!user.unEquip(O))
+			return
+		stock_item(O)
+		user.visible_message("<span class='notice'>\The [user] has added \the [O] to \the [src].</span>", "<span class='notice'>You add \the [O] to \the [src].</span>")
 
-			nanomanager.update_uis(src)
-
-	else if(istype(O, /obj/item/weapon/storage/bag))
+	else if(istype(O, /obj/item/weapon/storage))
 		var/obj/item/weapon/storage/bag/P = O
 		var/plants_loaded = 0
 		for(var/obj/G in P.contents)
-			if(accept_check(G))
-				if(contents.len >= max_n_of_items)
-					user << "<span class='notice'>\The [src] is full.</span>"
-					return 1
-				else
-					P.remove_from_storage(G,src)
-					if(item_quants[G.name])
-						item_quants[G.name]++
-					else
-						item_quants[G.name] = 1
-					plants_loaded++
+			if(accept_check(G) && P.remove_from_storage(G, src, 1))
+				plants_loaded++
+				stock_item(G)
+		P.finish_bulk_removal()
+
 		if(plants_loaded)
-
-			user.visible_message("<span class='notice'>[user] loads \the [src] with \the [P].</span>", "<span class='notice'>You load \the [src] with \the [P].</span>")
+			user.visible_message("<span class='notice'>\The [user] loads \the [src] with the contents of \the [P].</span>", "<span class='notice'>You load \the [src] with the contents of \the [P].</span>")
 			if(P.contents.len > 0)
-				user << "<span class='notice'>Some items are refused.</span>"
-
-		nanomanager.update_uis(src)
+				to_chat(user, "<span class='notice'>Some items were refused.</span>")
 
 	else
-		user << "<span class='notice'>\The [src] smartly refuses [O].</span>"
-		return 1
+		to_chat(user, "<span class='notice'>\The [src] smartly refuses [O].</span>")
+	return 1
 
-/obj/machinery/smartfridge/secure/attackby(var/obj/item/O as obj, var/mob/user as mob)
-	if(istype(O, /obj/item/weapon/card/emag))
+/obj/machinery/smartfridge/secure/emag_act(var/remaining_charges, var/mob/user)
+	if(!emagged)
 		emagged = 1
 		locked = -1
-		user << "You short out the product lock on [src]."
-		return
+		to_chat(user, "You short out the product lock on [src].")
+		return 1
 
-	..()
+/obj/machinery/smartfridge/proc/stock_item(var/obj/item/O)
+	for(var/datum/stored_items/I in item_records)
+		if(istype(O, I.item_path) && O.name == I.item_name)
+			stock(I, O)
+			return
+
+	var/datum/stored_items/I = new/datum/stored_items(src, O.type, O.name)
+	dd_insertObjectList(item_records, I)
+	stock(I, O)
+
+/obj/machinery/smartfridge/proc/stock(var/datum/stored_items/I, var/obj/item/O)
+	I.add_product(O)
+	SSnano.update_uis(src)
 
 /obj/machinery/smartfridge/attack_ai(mob/user as mob)
 	attack_hand(user)
@@ -269,16 +288,16 @@
 	data["secure"] = is_secure
 
 	var/list/items[0]
-	for (var/i=1 to length(item_quants))
-		var/K = item_quants[i]
-		var/count = item_quants[K]
+	for (var/i=1 to length(item_records))
+		var/datum/stored_items/I = item_records[i]
+		var/count = I.get_amount()
 		if(count > 0)
-			items.Add(list(list("display_name" = html_encode(capitalize(K)), "vend" = i, "quantity" = count)))
+			items.Add(list(list("display_name" = html_encode(capitalize(I.item_name)), "vend" = i, "quantity" = count)))
 
 	if(items.len > 0)
 		data["contents"] = items
 
-	ui = nanomanager.try_update_ui(user, src, ui_key, ui, data, force_open)
+	ui = SSnano.try_update_ui(user, src, ui_key, ui, data, force_open)
 	if(!ui)
 		ui = new(user, src, ui_key, "smartfridge.tmpl", src.name, 400, 500)
 		ui.set_initial_data(data)
@@ -288,9 +307,7 @@
 	if(..()) return 0
 
 	var/mob/user = usr
-	var/datum/nanoui/ui = nanomanager.get_open_ui(user, src, "main")
-
-	src.add_fingerprint(user)
+	var/datum/nanoui/ui = SSnano.get_open_ui(user, src, "main")
 
 	if(href_list["close"])
 		user.unset_machine()
@@ -300,20 +317,15 @@
 	if(href_list["vend"])
 		var/index = text2num(href_list["vend"])
 		var/amount = text2num(href_list["amount"])
-		var/K = item_quants[index]
-		var/count = item_quants[K]
+		var/datum/stored_items/I = item_records[index]
+		var/count = I.get_amount()
 
 		// Sanity check, there are probably ways to press the button when it shouldn't be possible.
 		if(count > 0)
-			item_quants[K] = max(count - amount, 0)
-
-			var/i = amount
-			for(var/obj/O in contents)
-				if(O.name == K)
-					O.loc = loc
-					i--
-					if(i <= 0)
-						return 1
+			if((count - amount) < 0)
+				amount = count
+			for(var/i = 1 to amount)
+				I.get_product(get_turf(src))
 
 		return 1
 	return 0
@@ -324,22 +336,17 @@
 	if(!target)
 		return 0
 
-	for (var/O in item_quants)
-		if(item_quants[O] <= 0) //Try to use a record that actually has something to dump.
+	for(var/datum/stored_items/I in src.item_records)
+		throw_item = I.get_product(loc)
+		if (!throw_item)
 			continue
-
-		item_quants[O]--
-		for(var/obj/T in contents)
-			if(T.name == O)
-				T.loc = src.loc
-				throw_item = T
-				break
 		break
+
 	if(!throw_item)
 		return 0
 	spawn(0)
 		throw_item.throw_at(target,16,3,src)
-	src.visible_message("\red <b>[src] launches [throw_item.name] at [target.name]!</b>")
+	src.visible_message("<span class='warning'>[src] launches [throw_item.name] at [target.name]!</span>")
 	return 1
 
 /************************
@@ -350,6 +357,6 @@
 	if(stat & (NOPOWER|BROKEN)) return 0
 	if(usr.contents.Find(src) || (in_range(src, usr) && istype(loc, /turf)))
 		if(!allowed(usr) && !emagged && locked != -1 && href_list["vend"])
-			usr << "<span class='warning'>Access denied.</span>"
+			to_chat(usr, "<span class='warning'>Access denied.</span>")
 			return 0
 	return ..()
