+++ code/game/machinery/pipe/pipe_dispenser.dm
@@ -17,6 +17,8 @@
 <A href='?src=\ref[src];make=1;dir=5'>Bent Pipe</A><BR>
 <A href='?src=\ref[src];make=5;dir=1'>Manifold</A><BR>
 <A href='?src=\ref[src];make=8;dir=1'>Manual Valve</A><BR>
+<A href='?src=\ref[src];make=9;dir=1'>Digital Valve</A><BR>
+<A href='?src=\ref[src];make=44;dir=1'>Automatic Shutoff Valve</A><BR>
 <A href='?src=\ref[src];make=20;dir=1'>Pipe Cap</A><BR>
 <A href='?src=\ref[src];make=19;dir=1'>4-Way Manifold</A><BR>
 <A href='?src=\ref[src];make=18;dir=1'>Manual T-Valve</A><BR>
@@ -39,14 +41,22 @@
 <A href='?src=\ref[src];make=36;dir=1'>4-Way Manifold</A><BR>
 <A href='?src=\ref[src];make=38;dir=1'>Upward Pipe</A><BR>
 <A href='?src=\ref[src];make=40;dir=1'>Downward Pipe</A><BR>
+<b>Fuel pipes:</b><BR>
+<A href='?src=\ref[src];make=45;dir=1'>Pipe</A><BR>
+<A href='?src=\ref[src];make=46;dir=5'>Bent Pipe</A><BR>
+<A href='?src=\ref[src];make=47;dir=1'>Manifold</A><BR>
+<A href='?src=\ref[src];make=51;dir=1'>Pipe Cap</A><BR>
+<A href='?src=\ref[src];make=48;dir=1'>4-Way Manifold</A><BR>
+<A href='?src=\ref[src];make=49;dir=1'>Upward Pipe</A><BR>
+<A href='?src=\ref[src];make=50;dir=1'>Downward Pipe</A><BR>
 <b>Devices:</b><BR>
 <A href='?src=\ref[src];make=28;dir=1'>Universal pipe adapter</A><BR>
 <A href='?src=\ref[src];make=4;dir=1'>Connector</A><BR>
 <A href='?src=\ref[src];make=7;dir=1'>Unary Vent</A><BR>
-<A href='?src=\ref[src];make=9;dir=1'>Gas Pump</A><BR>
+<A href='?src=\ref[src];make=10;dir=1'>Gas Pump</A><BR>
 <A href='?src=\ref[src];make=15;dir=1'>Pressure Regulator</A><BR>
 <A href='?src=\ref[src];make=16;dir=1'>High Power Gas Pump</A><BR>
-<A href='?src=\ref[src];make=10;dir=1'>Scrubber</A><BR>
+<A href='?src=\ref[src];make=11;dir=1'>Scrubber</A><BR>
 <A href='?src=\ref[src];makemeter=1'>Meter</A><BR>
 <A href='?src=\ref[src];make=13;dir=1'>Gas Filter</A><BR>
 <A href='?src=\ref[src];make=23;dir=1'>Gas Filter - Mirrored</A><BR>
@@ -60,9 +70,6 @@
 <A href='?src=\ref[src];make=3;dir=5'>Bent Pipe</A><BR>
 <A href='?src=\ref[src];make=6;dir=1'>Junction</A><BR>
 <A href='?src=\ref[src];make=17;dir=1'>Heat Exchanger</A><BR>
-<b>Insulated pipes:</b><BR>
-<A href='?src=\ref[src];make=11;dir=1'>Pipe</A><BR>
-<A href='?src=\ref[src];make=12;dir=5'>Bent Pipe</A><BR>
 
 "}
 ///// Z-Level stuff
@@ -72,21 +79,17 @@
 	onclose(user, "pipedispenser")
 	return
 
-/obj/machinery/pipedispenser/Topic(href, href_list)
-	if(..())
-		return
-	if(unwrenched || !usr.canmove || usr.stat || usr.restrained() || !in_range(loc, usr))
+/obj/machinery/pipedispenser/Topic(href, href_list, state = GLOB.physical_state)
+	if((. = ..()) || unwrenched)
 		usr << browse(null, "window=pipedispenser")
 		return
-	usr.set_machine(src)
-	src.add_fingerprint(usr)
+
 	if(href_list["make"])
 		if(!wait)
 			var/p_type = text2num(href_list["make"])
 			var/p_dir = text2num(href_list["dir"])
 			var/obj/item/pipe/P = new (/*usr.loc*/ src.loc, pipe_type=p_type, dir=p_dir)
 			P.update()
-			P.add_fingerprint(usr)
 			wait = 1
 			spawn(10)
 				wait = 0
@@ -96,23 +99,24 @@
 			wait = 1
 			spawn(15)
 				wait = 0
-	return
 
 /obj/machinery/pipedispenser/attackby(var/obj/item/W as obj, var/mob/user as mob)
-	src.add_fingerprint(usr)
 	if (istype(W, /obj/item/pipe) || istype(W, /obj/item/pipe_meter))
-		usr << "\blue You put [W] back to [src]."
-		user.drop_item()
+		if(!user.unEquip(W))
+			return
+		to_chat(usr, "<span class='notice'>You put \the [W] back into \the [src].</span>")
+		add_fingerprint(usr)
 		qdel(W)
 		return
-	else if (istype(W, /obj/item/weapon/wrench))
+	else if(isWrench(W))
+		add_fingerprint(usr)
 		if (unwrenched==0)
 			playsound(src.loc, 'sound/items/Ratchet.ogg', 50, 1)
-			user << "\blue You begin to unfasten \the [src] from the floor..."
-			if (do_after(user, 40))
+			to_chat(user, "<span class='notice'>You begin to unfasten \the [src] from the floor...</span>")
+			if (do_after(user, 40, src))
 				user.visible_message( \
-					"[user] unfastens \the [src].", \
-					"\blue You have unfastened \the [src]. Now it can be pulled somewhere else.", \
+					"<span class='notice'>\The [user] unfastens \the [src].</span>", \
+					"<span class='notice'>You have unfastened \the [src]. Now it can be pulled somewhere else.</span>", \
 					"You hear ratchet.")
 				src.anchored = 0
 				src.stat |= MAINT
@@ -121,11 +125,11 @@
 					usr << browse(null, "window=pipedispenser")
 		else /*if (unwrenched==1)*/
 			playsound(src.loc, 'sound/items/Ratchet.ogg', 50, 1)
-			user << "\blue You begin to fasten \the [src] to the floor..."
-			if (do_after(user, 20))
+			to_chat(user, "<span class='notice'>You begin to fasten \the [src] to the floor...</span>")
+			if (do_after(user, 20, src))
 				user.visible_message( \
-					"[user] fastens \the [src].", \
-					"\blue You have fastened \the [src]. Now it can dispense pipes.", \
+					"<span class='notice'>\The [user] fastens \the [src].</span>", \
+					"<span class='notice'>You have fastened \the [src]. Now it can dispense pipes.</span>", \
 					"You hear ratchet.")
 				src.anchored = 1
 				src.stat &= ~MAINT
@@ -151,11 +155,11 @@ Nah
 */
 
 //Allow you to drag-drop disposal pipes into it
-/obj/machinery/pipedispenser/disposal/MouseDrop_T(var/obj/structure/disposalconstruct/pipe as obj, mob/usr as mob)
-	if(!usr.canmove || usr.stat || usr.restrained())
+/obj/machinery/pipedispenser/disposal/MouseDrop_T(var/obj/structure/disposalconstruct/pipe as obj, mob/user as mob)
+	if(!CanPhysicallyInteract(user))
 		return
 
-	if (!istype(pipe) || get_dist(usr, src) > 1 || get_dist(src,pipe) > 1 )
+	if (!istype(pipe) || get_dist(src,pipe) > 1 )
 		return
 
 	if (pipe.anchored)
@@ -184,6 +188,8 @@ Nah
 <A href='?src=\ref[src];dmake=10'>Sorting (Untagged)</A><BR>
 <A href='?src=\ref[src];dmake=11'>Tagger</A><BR>
 <A href='?src=\ref[src];dmake=12'>Tagger (Partial)</A><BR>
+<A href='?src=\ref[src];dmake=13'>Diversion</A><BR>
+<A href='?src=\ref[src];dmake=14'>Diversion Switch</A><BR>
 "}
 ///// Z-Level stuff
 
@@ -193,59 +199,60 @@ Nah
 // 0=straight, 1=bent, 2=junction-j1, 3=junction-j2, 4=junction-y, 5=trunk
 
 
-/obj/machinery/pipedispenser/disposal/Topic(href, href_list)
-	if(..())
+/obj/machinery/pipedispenser/disposal/Topic(href, href_list, state = GLOB.physical_state)
+	if((. = ..()) || unwrenched)
+		usr << browse(null, "window=pipedispenser")
 		return
-	usr.set_machine(src)
-	src.add_fingerprint(usr)
+
 	if(href_list["dmake"])
-		if(unwrenched || !usr.canmove || usr.stat || usr.restrained() || !in_range(loc, usr))
-			usr << browse(null, "window=pipedispenser")
-			return
 		if(!wait)
 			var/p_type = text2num(href_list["dmake"])
-			var/obj/structure/disposalconstruct/C = new (src.loc)
-			switch(p_type)
-				if(0)
-					C.ptype = 0
-				if(1)
-					C.ptype = 1
-				if(2)
-					C.ptype = 2
-				if(3)
-					C.ptype = 4
-				if(4)
-					C.ptype = 5
-				if(5)
-					C.ptype = 6
-					C.density = 1
-				if(6)
-					C.ptype = 7
-					C.density = 1
-				if(7)
-					C.ptype = 8
-					C.density = 1
-				if(8)
-					C.ptype = 9
-					C.subtype = 0
-				if(9)
-					C.ptype = 9
-					C.subtype = 1
-				if(10)
-					C.ptype = 9
-					C.subtype = 2
-				if(11)
-					C.ptype = 13
-				if(12)
-					C.ptype = 14
+			if(p_type == 15)
+				new /obj/machinery/disposal_switch (get_turf(src))
+			else
+				var/obj/structure/disposalconstruct/C = new (src.loc)
+				switch(p_type)
+					if(0)
+						C.ptype = 0
+					if(1)
+						C.ptype = 1
+					if(2)
+						C.ptype = 2
+					if(3)
+						C.ptype = 4
+					if(4)
+						C.ptype = 5
+					if(5)
+						C.ptype = 6
+						C.set_density(1)
+					if(6)
+						C.ptype = 7
+						C.set_density(1)
+					if(7)
+						C.ptype = 8
+						C.set_density(1)
+					if(8)
+						C.ptype = 9
+						C.subtype = 0
+					if(9)
+						C.ptype = 9
+						C.subtype = 1
+					if(10)
+						C.ptype = 9
+						C.subtype = 2
+					if(11)
+						C.ptype = 13
+					if(12)
+						C.ptype = 14
+					if(13)
+						C.ptype = 15
 ///// Z-Level stuff
-				if(21)
-					C.ptype = 11
-				if(22)
-					C.ptype = 12
+					if(21)
+						C.ptype = 11
+					if(22)
+						C.ptype = 12
 ///// Z-Level stuff
-			C.add_fingerprint(usr)
-			C.update()
+				C.update()
 			wait = 1
 			spawn(15)
 				wait = 0
